[#55967] vsd: build: Adjust created board files
After transition to HWMv2 in Zephyr, the new file - board.yml - is
needed in the board directory. Additionally, the name of Kconfig.board
has changed to Kconfig.{board_name}. This name used used to automatically
match the name of board with the default configuration, hence we had to
change created name (replaced any dash or plus symbols wit underscore).
After changes in data generated for VSD app, we also have to add new
selected flag to the Kconfig file.