[#57989] demo: blinky-temperature: Update how led state is printed
The LED gpio pin isn't configured to support input. It caused problems,
because for some gpio drivers the read from pin not configured for
reading is different than we expect. We should keep the expected led
state in separate variable in our application (as it is already done in
Zephyr blinky sample).